+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/AnimalMetadata.java
@@ -8,8 +8,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class AnimalMetadata extends AgeableMobMetadata {
 
-    public static final byte OFFSET = AgeableMobMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 0;
+    public static final int OFFSET = AgeableMobMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 0;
 
     protected AnimalMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
