+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/HorseMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class HorseMetadata extends AbstractHorseMetadata {
 
-    public static final byte OFFSET = AbstractHorseMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 1;
+    public static final int OFFSET = AbstractHorseMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 1;
 
     public HorseMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
