+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/TurtleMetadata.java
@@ -8,8 +8,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class TurtleMetadata extends AnimalMetadata {
 
-    public static final byte OFFSET = AnimalMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 6;
+    public static final int OFFSET = AnimalMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 6;
 
     public TurtleMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
