+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/ZombieHorseMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class ZombieHorseMetadata extends AbstractHorseMetadata {
 
-    public static final byte OFFSET = AbstractHorseMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 0;
+    public static final int OFFSET = AbstractHorseMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 0;
 
     public ZombieHorseMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
