+++ src/main/java/mc/replay/wrapper/entity/metadata/minecart/CommandBlockMinecartMetadata.java
@@ -8,8 +8,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class CommandBlockMinecartMetadata extends AbstractMinecartMetadata {
 
-    public static final byte OFFSET = AbstractMinecartMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 2;
+    public static final int OFFSET = AbstractMinecartMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 2;
 
     public CommandBlockMinecartMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
